Modifying input param is a very bad idea
This fix PE-1366. Root cause is that when the resolve method is called twice (the 1st time is to check if it a SQL source so that we can enforce authz, the 2nd time is to actual load the DDF) the 2nd time it will receive an incorrect options value.